requirements.txt
async-timeout==4.0.2
hiredis==2.2.3
redis==4.6.0
python-dotenv==0.17.1